+++ spring-cloud-aws-autoconfigure/src/test/java/org/springframework/cloud/aws/autoconfigure/secretsmanager/AwsSecretsManagerBootstrapConfigurationTest.java
@@ -16,18 +16,30 @@
 
 package org.springframework.cloud.aws.autoconfigure.secretsmanager;
 
+import java.lang.reflect.Method;
+
+import com.amazonaws.AmazonWebServiceClient;
 import com.amazonaws.services.secretsmanager.AWSSecretsManager;
+import com.amazonaws.services.secretsmanager.AWSSecretsManagerClient;
 import org.junit.jupiter.api.Test;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.boot.test.context.runner.ApplicationContextRunner;
+import org.springframework.cloud.aws.secretsmanager.AwsSecretsManagerProperties;
 import org.springframework.cloud.aws.secretsmanager.AwsSecretsManagerPropertySourceLocator;
 import org.springframework.context.annotation.Bean;
+import org.springframework.util.ReflectionUtils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.mock;
 
+/**
+ * Test for {@link AwsSecretsManagerBootstrapConfiguration}.
+ *
+ * @author Matej Nedic
+ * @author Eddú Meléndez
+ */
 class AwsSecretsManagerBootstrapConfigurationTest {
 
 	private final ApplicationContextRunner contextRunner = new ApplicationContextRunner()
@@ -46,6 +58,30 @@ class AwsSecretsManagerBootstrapConfigurationTest {
 		this.contextRunner.withPropertyValues("aws.secretsmanager.enabled=false")
 				.run((context) -> assertThat(context)
 						.doesNotHaveBean(AwsSecretsManagerPropertySourceLocator.class));
+	}
+
+	@Test
+	void testWithStaticRegion() {
+		AwsSecretsManagerBootstrapConfiguration bootstrapConfig = new AwsSecretsManagerBootstrapConfiguration();
+
+		String region = "us-east-2";
+		AwsSecretsManagerProperties awsParamStoreProperties = new AwsSecretsManagerProperties();
+		awsParamStoreProperties.setRegion(region);
+
+		Method SMClientMethod = ReflectionUtils.findMethod(
+				AwsSecretsManagerBootstrapConfiguration.class, "smClient",
+				AwsSecretsManagerProperties.class);
+		SMClientMethod.setAccessible(true);
+		AWSSecretsManagerClient awsSimpleClient = (AWSSecretsManagerClient) ReflectionUtils
+				.invokeMethod(SMClientMethod, bootstrapConfig, awsParamStoreProperties);
+
+		Method signingRegionMethod = ReflectionUtils
+				.findMethod(AmazonWebServiceClient.class, "getSigningRegion");
+		signingRegionMethod.setAccessible(true);
+		String signedRegion = (String) ReflectionUtils.invokeMethod(signingRegionMethod,
+				awsSimpleClient);
+
+		assertThat(signedRegion).isEqualTo(region);
 	}
 
 	@TestConfiguration
